Share the draggable overlay behaviour instead of writing it twice
Sonar's duplication gate was pointing at something real: the pinned capture and
the window replica had the same frameless setup, the same drag handling and the
same double-click close, written out twice.

Both are things you refer to rather than backdrops, so both need that behaviour.
Kept separate, changing how one of them closes would quietly leave the other
behaving differently with nothing to explain why.

DraggableTopWindow holds it once and both inherit it.
